+++ cr/patch/symfony/polyfill/src/Php80/bootstrap.php
@@ -26,13 +26,13 @@ if (!function_exists('preg_last_error_msg')) {
     function preg_last_error_msg() { return p\Php80::preg_last_error_msg(); }
 }
 if (!function_exists('str_contains')) {
-    function str_contains($haystack, $needle) { return p\Php80::str_contains($haystack, $needle); }
+    function str_contains($haystack = null, $needle = null) { return p\Php80::str_contains(isset($haystack) ? $haystack : '', isset($needle) ? $needle : ''); }
 }
 if (!function_exists('str_starts_with')) {
-    function str_starts_with($haystack, $needle) { return p\Php80::str_starts_with($haystack, $needle); }
+    function str_starts_with($haystack = null, $needle = null) { return p\Php80::str_starts_with(isset($haystack) ? $haystack : '', isset($needle) ? $needle : ''); }
 }
 if (!function_exists('str_ends_with')) {
-    function str_ends_with($haystack, $needle) { return p\Php80::str_ends_with($haystack, $needle); }
+    function str_ends_with($haystack = null, $needle = null) { return p\Php80::str_ends_with(isset($haystack) ? $haystack : '', isset($needle) ? $needle : ''); }
 }
 if (!function_exists('get_debug_type')) {
     function get_debug_type($value) { return p\Php80::get_debug_type($value); }