+++ cr/symfony/error-handler/ThrowableUtils.php
@@ -11,14 +11,19 @@
 
 namespace Symfony\Component\ErrorHandler;
 
+use Symfony\Component\ErrorHandler\Exception\SilencedErrorContext;
+
 /**
  * @internal
  */
 class ThrowableUtils
 {
+    /**
+     * @param SilencedErrorContext|\Throwable
+     */
     public static function getSeverity($throwable) //// int
     {
-        if ($throwable instanceof \ErrorException) {
+        if ($throwable instanceof \ErrorException || $throwable instanceof SilencedErrorContext) {
             return $throwable->getSeverity();
         }