+++ cr/symfony/mime/Crypto/SMime.php
@@ -93,7 +93,7 @@ abstract class SMime
             }
 
             // Handle headers that span multiple lines
-            if (false === strpos($headerLine, ':')) {
+            if (!str_contains($headerLine, ':')) {
                 $headers[$currentHeaderName] .= ' '.trim($headerLine);
                 continue;
             }