+++ cr/symfony/mime/Part/DataPart.php
@@ -45,8 +45,10 @@ class DataPart extends TextPart
 
         parent::__construct($body, null, $subtype, $encoding);
 
-        $this->filename = $filename;
-        $this->setName($filename);
+        if (null !== $filename) {
+            $this->filename = $filename;
+            $this->setName($filename);
+        }
         $this->setDisposition('attachment');
     }