+++ cr/symfony/routing/Route.php
@@ -118,8 +118,6 @@ class Route implements \Serializable
     }
 
     /**
-     * Returns the pattern for the path.
-     *
      * @return string The path pattern
      */
     public function getPath()
@@ -128,10 +126,6 @@ class Route implements \Serializable
     }
 
     /**
-     * Sets the pattern for the path.
-     *
-     * This method implements a fluent interface.
-     *
      * @return $this
      */
     public function setPath($pattern)
@@ -149,8 +143,6 @@ class Route implements \Serializable
     }
 
     /**
-     * Returns the pattern for the host.
-     *
      * @return string The host pattern
      */
     public function getHost()
@@ -159,10 +151,6 @@ class Route implements \Serializable
     }
 
     /**
-     * Sets the pattern for the host.
-     *
-     * This method implements a fluent interface.
-     *
      * @return $this
      */
     public function setHost($pattern = null)
@@ -189,8 +177,6 @@ class Route implements \Serializable
     /**
      * Sets the schemes (e.g. 'https') this route is restricted to.
      * So an empty array means that any scheme is allowed.
-     *
-     * This method implements a fluent interface.
      *
      * @param string|string[] $schemes The scheme or an array of schemes
      *
@@ -231,8 +217,6 @@ class Route implements \Serializable
      * Sets the HTTP methods (e.g. 'POST') this route is restricted to.
      * So an empty array means that any method is allowed.
      *
-     * This method implements a fluent interface.
-     *
      * @param string|string[] $methods The method or an array of methods
      *
      * @return $this
@@ -246,8 +230,6 @@ class Route implements \Serializable
     }
 
     /**
-     * Returns the options.
-     *
      * @return array The options
      */
     public function getOptions()
@@ -256,10 +238,6 @@ class Route implements \Serializable
     }
 
     /**
-     * Sets the options.
-     *
-     * This method implements a fluent interface.
-     *
      * @return $this
      */
     public function setOptions(array $options)
@@ -272,10 +250,6 @@ class Route implements \Serializable
     }
 
     /**
-     * Adds options.
-     *
-     * This method implements a fluent interface.
-     *
      * @return $this
      */
     public function addOptions(array $options)
@@ -290,8 +264,6 @@ class Route implements \Serializable
 
     /**
      * Sets an option value.
-     *
-     * This method implements a fluent interface.
      *
      * @param mixed $value The option value
      *
@@ -308,8 +280,6 @@ class Route implements \Serializable
     }
 
     /**
-     * Get an option value.
-     *
      * @return mixed The option value or null when not given
      */
     public function getOption($name)
@@ -320,8 +290,6 @@ class Route implements \Serializable
     }
 
     /**
-     * Checks if an option has been set.
-     *
      * @return bool true if the option is set, false otherwise
      */
     public function hasOption($name)
@@ -332,8 +300,6 @@ class Route implements \Serializable
     }
 
     /**
-     * Returns the defaults.
-     *
      * @return array The defaults
      */
     public function getDefaults()
@@ -342,12 +308,6 @@ class Route implements \Serializable
     }
 
     /**
-     * Sets the defaults.
-     *
-     * This method implements a fluent interface.
-     *
-     * @param array $defaults The defaults
-     *
      * @return $this
      */
     public function setDefaults(array $defaults)
@@ -358,12 +318,6 @@ class Route implements \Serializable
     }
 
     /**
-     * Adds defaults.
-     *
-     * This method implements a fluent interface.
-     *
-     * @param array $defaults The defaults
-     *
      * @return $this
      */
     public function addDefaults(array $defaults)
@@ -381,8 +335,6 @@ class Route implements \Serializable
     }
 
     /**
-     * Gets a default value.
-     *
      * @return mixed The default value or null when not given
      */
     public function getDefault($name)
@@ -393,8 +345,6 @@ class Route implements \Serializable
     }
 
     /**
-     * Checks if a default value is set for the given variable.
-     *
      * @return bool true if the default value is set, false otherwise
      */
     public function hasDefault($name)
@@ -426,8 +376,6 @@ class Route implements \Serializable
     }
 
     /**
-     * Returns the requirements.
-     *
      * @return array The requirements
      */
     public function getRequirements()
@@ -436,12 +384,6 @@ class Route implements \Serializable
     }
 
     /**
-     * Sets the requirements.
-     *
-     * This method implements a fluent interface.
-     *
-     * @param array $requirements The requirements
-     *
      * @return $this
      */
     public function setRequirements(array $requirements)
@@ -452,12 +394,6 @@ class Route implements \Serializable
     }
 
     /**
-     * Adds requirements.
-     *
-     * This method implements a fluent interface.
-     *
-     * @param array $requirements The requirements
-     *
      * @return $this
      */
     public function addRequirements(array $requirements)
@@ -475,8 +411,6 @@ class Route implements \Serializable
     }
 
     /**
-     * Returns the requirement for the given key.
-     *
      * @return string|null The regex or null when not given
      */
     public function getRequirement($key)
@@ -487,8 +421,6 @@ class Route implements \Serializable
     }
 
     /**
-     * Checks if a requirement is set for the given key.
-     *
      * @return bool true if a requirement is specified, false otherwise
      */
     public function hasRequirement($key)
@@ -499,8 +431,6 @@ class Route implements \Serializable
     }
 
     /**
-     * Sets a requirement for the given key.
-     *
      * @return $this
      */
     public function setRequirement($key, $regex)
@@ -520,8 +450,6 @@ class Route implements \Serializable
     }
 
     /**
-     * Returns the condition.
-     *
      * @return string The condition
      */
     public function getCondition()
@@ -530,10 +458,6 @@ class Route implements \Serializable
     }
 
     /**
-     * Sets the condition.
-     *
-     * This method implements a fluent interface.
-     *
      * @return $this
      */
     public function setCondition($condition = null)
@@ -601,7 +525,7 @@ class Route implements \Serializable
             }
         }
 
-        if ('$' === substr($regex, -1)) {
+        if (str_ends_with($regex, '$')) {
             $regex = substr($regex, 0, -1);
         } elseif (\strlen($regex) - 2 === strpos($regex, '\\z')) {
             $regex = substr($regex, 0, -2);